requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
arrow==1.0.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
diff-match-patch==20200713
feedparser==6.0.2
furl==2.1.2
idna==2.10
imgkit==1.2.2
multidict==5.1.0
mypy-extensions==0.4.3
orderedmultidict==1.0.1
pathspec==0.8.1
pony==0.7.14
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.17.0
regex==2021.4.4
requests==2.25.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2.1
telepot==12.7
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
